converted purchase/selling event units to double

+++ Resourcerer_Sol/src/Resourcerer.Dtos/Elements/ElementUsageDetailsDto.cs
@@ -5,7 +5,7 @@ public class ElementUsageDetailsDto : BaseDto
     public Guid ElementId { get; set; }
     public string? ElementName { get; set; }
     public string? Unit { get; set; }
-    public int UnitsPurchased { get; set; }
+    public double UnitsPurchased { get; set; }
     public double PurchaseCosts { get; set; }
     public double UnitsUsed { get; set; }
     public double UnitsInStock { get; set; }
